add child message count

+++ vonline_be/src/com/vmesteonline/be/jdo2/VoMessage.java
@@ -176,7 +176,7 @@ public class VoMessage extends VoBaseMessage {
 
 	protected int visibleOffset;
 
-	public long getVisibleOffset() {
+	public int getVisibleOffset() {
 		return visibleOffset;
 	}
